Refactor payment event handling: move events to domain package and update imports
