+++ payment-domain/src/main/kotlin/com/dogancaglar/payment/domain/OutboxEvent.kt
