+++ payment-domain/src/main/kotlin/com/dogancaglar/payment/domain/PaymentOrderEvent.kt
@@ -1,4 +1,4 @@
-package com.dogancaglar.payment.application.events
+package com.dogancaglar.payment.domain
 
 import java.math.BigDecimal
 import java.time.LocalDateTime
