+++ payment-domain/src/main/kotlin/com/dogancaglar/payment/domain/ScheduledPaymentOrderStatusRequest.kt
@@ -1,4 +1,4 @@
-package com.dogancaglar.application
+package com.dogancaglar.payment.domain
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
