+++ payment-domain/src/main/kotlin/com/dogancaglar/payment/domain/factory/PaymentOrderFactory.kt
@@ -1,6 +1,6 @@
 package com.dogancaglar.payment.domain.factory
 
-import com.dogancaglar.payment.application.events.PaymentOrderEvent
+import com.dogancaglar.payment.domain.PaymentOrderEvent
 import com.dogancaglar.payment.domain.model.Amount
 import com.dogancaglar.payment.domain.model.PaymentOrder
 import com.dogancaglar.payment.domain.model.PaymentOrderStatus
@@ -9,8 +9,8 @@ import com.dogancaglar.payment.domain.model.vo.PaymentOrderId
 import com.dogancaglar.payment.domain.model.vo.SellerId
 
 class PaymentOrderFactory {
-    fun fromEvent(event: PaymentOrderEvent): PaymentOrder =
-        PaymentOrder.reconstructFromPersistence(
+    fun fromEvent(event: PaymentOrderEvent): PaymentOrder {
+        return PaymentOrder(
             paymentOrderId = PaymentOrderId(event.paymentOrderId.toLong()),
             publicPaymentOrderId = event.publicPaymentOrderId,
             paymentId = PaymentId(event.paymentId.toLong()),
@@ -24,4 +24,5 @@ class PaymentOrderFactory {
             retryReason = event.retryReason,
             lastErrorMessage = event.lastErrorMessage
         )
+    }
 }
