+++ payment-infrastructure/src/main/kotlin/com/dogancaglar/infrastructure/adapter/persistance/PaymentOutboundAdapter.kt
@@ -2,9 +2,9 @@ package com.dogancaglar.infrastructure.adapter.persistance
 
 import com.dogancaglar.infrastructure.mapper.PaymentEntityMapper
 import com.dogancaglar.infrastructure.persistence.repository.PaymentMapper
+import com.dogancaglar.payment.application.port.outbound.PaymentRepository
 import com.dogancaglar.payment.domain.model.Payment
 import com.dogancaglar.payment.domain.model.vo.PaymentId
-import com.dogancaglar.payment.domain.port.PaymentRepository
 import org.springframework.stereotype.Repository
 
 @Repository
