+++ payment-infrastructure/src/main/kotlin/com/dogancaglar/infrastructure/config/kafka/deserialization/EventEnvelopeDeserializer.kt
@@ -1,7 +1,7 @@
 package com.dogancaglar.infrastructure.config.kafka.deserialization
 
 import com.dogancaglar.common.event.EventEnvelope
-import com.dogancaglar.payment.application.events.EventMetadatas
+import com.dogancaglar.payment.domain.model.EventMetadatas
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
