+++ payment-infrastructure/src/main/kotlin/com/dogancaglar/infrastructure/redis/PaymentRetryQueueAdapter.kt
@@ -1,12 +1,12 @@
 package com.dogancaglar.infrastructure.redis
 
-import com.dogancaglar.application.PaymentOrderRetryRequested
 import com.dogancaglar.common.event.DomainEventEnvelopeFactory
 import com.dogancaglar.common.event.EventEnvelope
 import com.dogancaglar.common.logging.LogContext
-import com.dogancaglar.payment.application.events.EventMetadatas
 import com.dogancaglar.payment.application.mapper.PaymentOrderDomainEventMapper
 import com.dogancaglar.payment.application.port.outbound.RetryQueuePort
+import com.dogancaglar.payment.domain.PaymentOrderRetryRequested
+import com.dogancaglar.payment.domain.model.EventMetadatas
 import com.dogancaglar.payment.domain.model.PaymentOrder
 import com.dogancaglar.payment.domain.model.vo.PaymentOrderId
 import com.fasterxml.jackson.core.type.TypeReference
