+++ payment-service/src/test/kotlin/com/dogancaglar/domain/event/DomainEventFactoryTest.kt
@@ -1,10 +1,10 @@
 package com.dogancaglar.domain.event
 
-import com.dogancaglar.application.PaymentOrderCreated
 import com.dogancaglar.common.event.DomainEventEnvelopeFactory
 import com.dogancaglar.common.event.EventEnvelope
 import com.dogancaglar.common.logging.GenericLogFields
-import com.dogancaglar.payment.application.events.EventMetadatas
+import com.dogancaglar.payment.domain.PaymentOrderCreated
+import com.dogancaglar.payment.domain.model.EventMetadatas
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
